+++ native_library/ios/GoogleMapsANE/GoogleMapsANE_FW/MapKit/CustomMKPolygon.swift
@@ -20,7 +20,7 @@ import MapKit
 import UIKit
 
 internal class CustomMKPolygon: MKPolygon {
-    var identifier: String = ""
+    var identifier = ""
     var strokeWidth: CGFloat = 1.0
     var strokeColor: UIColor?
     var fillColor: UIColor?